Relative control of gravity.

+++ SUN/Include/ScreenSize.h
@@ -6,4 +6,4 @@ using namespace gmtl;
 static Vec2f screenSize(800, 600);
 static Vec2f quadPos(0, 0);
 //static Vec2f quadSize(1.0, 1.0);
-static Vec2f boundarySize(0.8, 0.8);
+static Vec2d boundarySize(0.8, 0.8);
